Fix for bug 217.
This patch modifies execute script (ddlscript_complete_int) so that it will
reconfigure the trigger arguments on the log trigger for any tables that need it.

This version of the patch only reconfigures tables already locked by the SQL script,
but the repair_log_triggers() function provides can be called with false to
repair all tables.